refactor(util): k8s helper methods
